Implement fill options to plotrs
Use the fillrange option from Plots to fill traces
either above or below the trace zero in record
sections, optionally allowing the level to be
set as a fraction of the maximum trace amplitude.

This does not draw the region properly, as we rely
on inserting NaNs at points above/below the threshold
value, and hence the filled region is missing some
area.  However, it's good enough when the time
range of the plots is much larger than the period
of the wave.